refactor: rename 'refuel' page to 'manage-pool'
