+++ apps/main/src/dex/features/manage-pool/components/BorrowInformationContainer.tsx
