+++ apps/main/src/dex/features/manage-pool/components/Header.tsx
