+++ apps/main/src/dex/features/manage-pool/components/ReservesCompositionChart.tsx
